requirements.txt
discord.py[voice]
asyncio
peony-twitter[all]
markovify
requests
configparser
nltk
networkx
numpy